Upgrade dependencies and refactor package name.
