+++ devopsguru_eks_test/src/main/java/com/amazonaws/devopsguru/eks/test/GatewayException.java
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT-0
-package com.amazonaws.capstone.devopsguru_eks_test;
+package com.amazonaws.devopsguru.eks.test;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
