+++ devopsguru_eks_test/src/test/java/com/amazonaws/devopsguru/eks/test/DevopsguruEksTestApplicationTests.java
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT-0
-package com.amazonaws.capstone.devopsguru_eks_test;
+package com.amazonaws.devopsguru.eks.test;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
